Define a Peer abstraction
