+++ lib/distributed-peer.js
@@ -1,13 +1,11 @@
-const EventEmitter = require('events')
+const Peer = require('./peer.js')
 const Messages = require('./messages.js')
 const MessageFactory = require('./message-factory.js')
-const debug = require('debug')('slsk:distributed-peer:i')
+const debug = require('debug')('slsk:peer:distributed:i')
 
-module.exports = class DistributedPeer extends EventEmitter {
+module.exports = class DistributedPeer extends Peer {
   constructor (socket, peer) {
-    super()
-    this.conn = socket
-    this.peer = peer
+    super(socket, peer)
 
     this.conn.on('connect', () => {
       if (peer.token) {
@@ -18,16 +16,6 @@ module.exports = class DistributedPeer extends EventEmitter {
         let buf = Buffer.from('05000000' + '00' + peer.token, 'hex')
         this.conn.write(buf)
       }
-    })
-
-    this.conn.on('error', error => {
-      debug(`${peer.user} error ${error.code}`)
-      this.emit('disconnect', {})
-    })
-
-    this.conn.on('end', () => {
-      debug(`${peer.user} connection ended`)
-      this.emit('disconnect', {})
     })
 
     let msgs = new Messages()
@@ -65,23 +53,5 @@ module.exports = class DistributedPeer extends EventEmitter {
         }
       }
     })
-  }
-
-  transferRequest (file, token) {
-    debug(`Transfer request ${file}`)
-    this.conn.write(MessageFactory
-      .to.peer
-      .transferRequest(file, token)
-      .getBuff())
-  }
-
-  setAddress (host, port) {
-    debug(`setAddress for ${this.peer.user}: ${host} ${port}`)
-    this.peer.host = host
-    this.peer.port = port
-  }
-
-  destroy () {
-    this.conn.destroy()
   }
 }
