+++ lib/slsk-client.js
@@ -7,7 +7,7 @@ const debug = require('debug')('slsk:i')
 
 const downloadPeerFile = require('./download-peer-file.js')
 const Server = require('./server.js')
-const Peer = require('./peer.js')
+const DefaultPeer = require('./default-peer.js')
 const DistributedPeer = require('./distributed-peer.js')
 const Listen = require('./listen.js')
 let stack = require('./stack')
@@ -54,7 +54,7 @@ class SlskClient extends EventEmitter {
         break
       }
       default: {
-        peers[peer.user] = new Peer(net.createConnection({
+        peers[peer.user] = new DefaultPeer(net.createConnection({
           host: peer.host,
           port: peer.port
         }), peer)
@@ -77,7 +77,7 @@ class SlskClient extends EventEmitter {
       } else {
         server.getPeerAddress(peer.user)
         debug(`new Peer connected ${peer.user} token ${peer.token}`)
-        peers[peer.user] = new Peer(evt.socket, peer)
+        peers[peer.user] = new DefaultPeer(evt.socket, peer)
         peers[peer.user].on('disconnect', () => {
           delete peers[peer.user]
         })
